MFC: r317578
Fix a bug introduced as part of r287726 (MFCed to stable/10 in
r292789); use the right device_t for determining the softc of
the bridge in psycho_route_interrupt(). [1]
While at it, update the corresponding comment that the code in
question is also necessary for U30s in addition to E450s (a fact
that has been known for ages).

PR:		218478 [1]
Submitted by:	Yoshihiko Iwama [1]