Upgrade OpenSSH to 7.3p1.
This is the last version of OpenSSH which does not break compatibility
more than we can live with in a stable branch.  Further commits will
follow to backport some bug fixes from newer versions.

The sshd breakage in the previous attempt was due to an upstream bug
(a 0 was changed to a 1 while refactoring send_rexec_state() in sshd.c)
which only manifested itself when sshd was built with SSH 1 support.

Approved by:	re@